added watcher controller

+++ service/db/type.go
@@ -21,7 +21,7 @@ type (
 	}
 
 	Behaviour struct {
-		ID        uint           `gorm:"primaryKey;autoIncrement" json:"id"`
+		ID        uint64         `gorm:"primaryKey;autoIncrement" json:"id"`
 		Uid       int64          `json:"uid" gorm:"index"`
 		CreatedAt time.Time      `json:"created_at" gorm:"index"`
 		TargetUid int64          `json:"target_uid" gorm:"index"`
@@ -55,7 +55,7 @@ type (
 	}
 
 	WatcherBehaviour struct {
-		ID        uint           `gorm:"primaryKey;autoIncrement" json:"id"`
+		ID        uint64         `gorm:"primaryKey;autoIncrement" json:"id"`
 		Uid       int64          `json:"uid" gorm:"index"`
 		UName     string         `json:"u_name"`
 		CreatedAt time.Time      `json:"created_at" gorm:"index"`
